fix nomad
