+++ Games/types/Mafia/const/Priority.js
@@ -25,6 +25,7 @@ const PRIORITY_BITING_WOLF = -50
 
 const PRIORITY_MODIFY_INVESTIGATIVE_RESULT_DEFAULT = -50
 const PRIORITY_REVEAL_TARGET_ON_DEATH = PRIORITY_MODIFY_INVESTIGATIVE_RESULT_DEFAULT
+const PRIORITY_MODIFY_ALIGNMENT = PRIORITY_MODIFY_INVESTIGATIVE_RESULT_DEFAULT
 
 const PRIORITY_ITEM_GIVER_DEFAULT = -50
 const PRIORITY_EFFECT_GIVER_DEFAULT = -50
@@ -42,7 +43,6 @@ const PRIORITY_MIMIC_ROLE = -5
 const PRIORITY_SWAP_ROLES = -4
 const PRIORITY_CLEAN_DEATH = -3
 const PRIORITY_IDENTITY_STEALER = -2
-const PRIORITY_GAIN_ALIGNMENT = -2
 
 const PRIORITY_CULT_CONVERT = -2
 const PRIORITY_MASON_CONVERT = -2
@@ -100,6 +100,7 @@ module.exports = {
 
     PRIORITY_MODIFY_INVESTIGATIVE_RESULT_DEFAULT,
     PRIORITY_REVEAL_TARGET_ON_DEATH,
+    PRIORITY_MODIFY_ALIGNMENT,
 
     PRIORITY_ITEM_GIVER_DEFAULT,
     PRIORITY_EFFECT_GIVER_DEFAULT,
@@ -117,7 +118,6 @@ module.exports = {
     PRIORITY_SWAP_ROLES,
     PRIORITY_CLEAN_DEATH,
     PRIORITY_IDENTITY_STEALER,
-    PRIORITY_GAIN_ALIGNMENT,
 
     PRIORITY_CULT_CONVERT,
     PRIORITY_MASON_CONVERT,
