+++ Games/types/Mafia/roles/cards/ModifyAlignment.js
@@ -1,0 +1,33 @@
+const Card = require("../../Card");
+const { PRIORITY_MODIFY_ALIGNMENT } = require("../../const/Priority");
+
+module.exports = class ModifyAlignment extends Card {
+
+    constructor(role) {
+        super(role);
+        
+        this.winCount = "Village";
+
+        this.meetings = {
+            "Follow the ways of": {
+                states: ["Night"],
+                flags: ["voting"],
+                action: {
+                    priority: PRIORITY_MODIFY_ALIGNMENT,
+                    run: function () {
+                        let alignment = this.target.role.alignment;
+                        if (alignment == "Independent") {
+                            this.actor.queueAlert(`You follow ${this.target.name} but did not find a home.`)
+                            return;
+                        }
+
+                        this.actor.role.alignment = alignment;
+                        this.actor.queueAlert(`You follow ${this.target.name} and learn the ways of the ${alignment}.`);
+                    }
+                }
+            }
+        }
+        
+    }
+
+}
